+++ out/views/visualizer.js
@@ -1,7 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.VisualizerPanel = void 0;
-class VisualizerPanel {
-}
-exports.VisualizerPanel = VisualizerPanel;
-//# sourceMappingURL=visualizer.js.map